requirements.txt
appdirs==1.4.3
attrs==19.3.0
azure-cognitiveservices-vision-computervision==0.5.0
azure-common==1.1.23
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.9
idna==2.8
isodate==0.6.0
mccabe==0.6.1
msrest==0.6.10
oauthlib==3.1.0
pathspec==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.10.3
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.7
